pmie scripts: rework
- use _pstree_all() and/or _pstree_oneline() instead of pstree(1)
- support for -l /dev/tty to dodge log files, exec, .prev juggling,
  etc

modified:   src/pmie/pmie_check.sh
modified:   src/pmie/pmie_daily.sh
modified:   src/pmie/rc_pmie

+++ src/pmie/pmie_daily.sh
@@ -54,7 +54,10 @@ _cleanup()
 {
     $USE_SYSLOG && [ $status -ne 0 ] && \
     $PCP_SYSLOG_PROG -p daemon.error "$prog failed - see $PROGLOG"
-    [ -s "$PROGLOG" ] || rm -f "$PROGLOG"
+    if [ "$PROGLOG" != "/dev/tty" ]
+    then
+	[ -s "$PROGLOG" ] || rm -f "$PROGLOG"
+    fi
     lockfile=`cat $tmp/lock 2>/dev/null`
     [ -n "$lockfile" ] && rm -f "$lockfile"
     rm -rf $tmp
@@ -216,9 +219,16 @@ done
 if $SHOWME
 then
     :
+elif [ "$PROGLOG" = "/dev/tty" ]
+then
+    # special case for debugging ... no salt away previous, no chown, no exec
+    #
+    :
 else
     # Salt away previous log, if any ...
     #
+    PROGLOGDIR=`dirname "$PROGLOG"`
+    [ -d "$PROGLOGDIR" ] || mkdir_and_chown "$PROGLOGDIR" 755 $PCP_USER:$PCP_GROUP 2>/dev/null
     _save_prev_file "$PROGLOG"
     # After argument checking, everything must be logged to ensure no mail is
     # accidentally sent from cron.  Close stdout and stderr, then open stdout
